+++ module/data/character.mjs
@@ -42,7 +42,7 @@ export default class Knave2eCharacter extends Knave2eActorType {
             max: new fields.NumberField({ ...requiredInteger, initial: 0, min: 0 })
         });
         schema.wounds = new fields.SchemaField({
-            value: new fields.NumberField({ ...requiredInteger, initial: 0, min: 0 }),
+            value: new fields.NumberField({ ...requiredInteger, initial: 10, min: 0 }),
             max: new fields.NumberField({ ...requiredInteger, initial: 10, min: 0 }),
             progress: new fields.NumberField({ ...requiredInteger, initial: 100, min: 0 })
         });