+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+## 0.2.0
+- Added support for Power Attacks, with automatic weapon breaking
+- Visual update to character, recruit, and monster sheets
+- Re-colored HP and Wounds to match default token attribute bars in Foundry
+- Added in-line damage buttons to attack roll chat messages 
+- Active relics automatically roll with their description, and can print spell effects directly from chat
+- Reversed wound progression direction for more clear interaction with default token attribute bars
+- Simplified "number appearing" fields to accept a single string of type Dng(Wild), (e.g. 1d6(3d6))
+- Converted multiple notifications into single-user dialog popups to reduce chat spam
+
 ## 0.1.2
 - Added automatic reminder for maneuevers on an attack roll >= 21
 
@@ -9,5 +19,4 @@
 - Changed minor sheet formatting issues - only small changes now, as sheets will change drastically in a future release.
 
 ## 0.1.0
-
 - Initial release